backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
